refactor: Move constrained transfer classes to the new package structure(code)
