+++ application/src/main/java/org/opentripplanner/routing/algorithm/transferoptimization/services/TransferGenerator.java
@@ -16,6 +16,7 @@ import org.opentripplanner.routing.algorithm.transferoptimization.model.StopTime
 import org.opentripplanner.routing.algorithm.transferoptimization.model.TripStopTime;
 import org.opentripplanner.routing.algorithm.transferoptimization.model.TripToTripTransfer;
 import org.opentripplanner.transfer.constrained.model.ConstrainedTransfer;
+import org.opentripplanner.transfer.constrained.raptoradaptor.ConstrainedBoardingSearch;
 
 /**
  * This class is responsible for finding all possible transfers between each pair of transit legs
@@ -193,8 +194,7 @@ public class TransferGenerator<T extends RaptorTripSchedule> {
   }
 
   /**
-   * This code duplicates the logic in
-   * {@link org.opentripplanner.routing.algorithm.raptoradapter.transit.constrainedtransfer.ConstrainedBoardingSearch},
+   * This code duplicates the logic in {@link ConstrainedBoardingSearch},
    * see the {@code findTimetableTripInfo(RaptorTimeTable, Iterable, int, int, int)}) method.
    */
   private int calculateEarliestBoardTime(
