+++ application/src/main/java/org/opentripplanner/transfer/constrained/raptoradaptor/ConstrainedBoardingSearch.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.algorithm.raptoradapter.transit.constrainedtransfer;
+package org.opentripplanner.transfer.constrained.raptoradaptor;
 
 import java.util.LinkedList;
 import java.util.List;
