+++ application/src/main/java/org/opentripplanner/transfer/constrained/raptoradaptor/TransferForPattern.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.algorithm.raptoradapter.transit.constrainedtransfer;
+package org.opentripplanner.transfer.constrained.raptoradaptor;
 
 import javax.annotation.Nullable;
 import org.opentripplanner.raptor.api.model.RaptorTransferConstraint;
