+++ application/src/main/java/org/opentripplanner/transfer/constrained/raptoradaptor/TransferIndexGenerator.java
@@ -1,6 +1,4 @@
-package org.opentripplanner.routing.algorithm.raptoradapter.transit.constrainedtransfer;
-
-import static org.opentripplanner.routing.algorithm.raptoradapter.transit.constrainedtransfer.TransferPointForPatternFactory.createTransferPointForPattern;
+package org.opentripplanner.transfer.constrained.raptoradaptor;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -154,7 +152,7 @@ public class TransferIndexGenerator {
       return List.of();
     }
 
-    var sourcePoint = createTransferPointForPattern(station);
+    var sourcePoint = TransferPointForPatternFactory.createTransferPointForPattern(station);
     var result = new ArrayList<TPoint>();
 
     for (RoutingTripPattern pattern : patterns) {
@@ -176,7 +174,7 @@ public class TransferIndexGenerator {
       return List.of();
     }
 
-    var sourcePoint = createTransferPointForPattern(stop.getIndex());
+    var sourcePoint = TransferPointForPatternFactory.createTransferPointForPattern(stop.getIndex());
     var result = new ArrayList<TPoint>();
 
     for (RoutingTripPattern pattern : patterns) {
@@ -226,7 +224,10 @@ public class TransferIndexGenerator {
       // stopPosInPattern == -1 means stop is not on pattern
       if (stopPosInPattern >= 0) {
         int stopIndex = pattern.stopIndex(stopPosInPattern);
-        var sourcePoint = createTransferPointForPattern(route, stopIndex);
+        var sourcePoint = TransferPointForPatternFactory.createTransferPointForPattern(
+          route,
+          stopIndex
+        );
         points.add(new TPoint(pattern, sourcePoint, null, stopPosInPattern));
       }
     }
@@ -259,7 +260,7 @@ public class TransferIndexGenerator {
     int stopPosInPattern = point.getStopPositionInPattern();
     var scheduledPattern = scheduledPatterns.get(0);
     int stopIndex = scheduledPattern.stopIndex(stopPosInPattern);
-    var sourcePoint = createTransferPointForPattern(trip, stopIndex);
+    var sourcePoint = TransferPointForPatternFactory.createTransferPointForPattern(trip, stopIndex);
     TPoint scheduledPoint = new TPoint(scheduledPattern, sourcePoint, trip, stopPosInPattern);
 
     // Return early if only scheduled pattern exists
@@ -281,7 +282,10 @@ public class TransferIndexGenerator {
           res.add(
             new TPoint(
               pattern,
-              createTransferPointForPattern(trip, pattern.stopIndex(stopPosInPattern)),
+              TransferPointForPatternFactory.createTransferPointForPattern(
+                trip,
+                pattern.stopIndex(stopPosInPattern)
+              ),
               trip,
               stopPosInPattern
             )
