+++ application/src/main/java/org/opentripplanner/transfer/constrained/raptoradaptor/TransferPointMatcher.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.algorithm.raptoradapter.transit.constrainedtransfer;
+package org.opentripplanner.transfer.constrained.raptoradaptor;
 
 import org.opentripplanner.transit.model.timetable.Trip;
